+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kori-ui/svelte
 
+## 0.0.2-alpha.2
+
+### Patch Changes
+
+- unify components into svelte package and cleanup individual packages
+- Updated dependencies
+  - @kori-ui/utilities@0.0.2-alpha.2
+  - @kori-ui/styles@0.0.2-alpha.2
+
 ## 0.0.2-alpha.0
 
 ### Patch Changes